use component CountryDetails
